refactor: extract internal/service/kg/ sub-package
- Move Pipeline, scoring, search, types, helpers to internal/service/kg/
- Keep backward-compat type aliases and function wrappers in service package
- Update handler to import kg package directly
- Fix test package boundaries (consistent package kg)
