+++ internal/service/kg/pipeline.go
@@ -14,7 +14,7 @@
 //  limitations under the License.
 //
 
-package service
+package kg
 
 import (
 	"context"
