+++ internal/service/kg/testutil_test.go
@@ -1,0 +1,3 @@
+package kg
+
+func strPtr(s string) *string { return &s }
